Moving game scene to game scene package
