+++ ForceGamaAttack/src/scenes/menu/MenuScene.java
@@ -1,6 +1,6 @@
 package scenes.menu;
 import scenes.Scene;
-import scenes.GameScene;
+import scenes.game.GameScene;
 import jplay.GameImage;
 import jplay.Keyboard;
 import java.util.ArrayList;
@@ -22,7 +22,7 @@ public class MenuScene extends Scene {
 	protected void initialSetup(){
 		
 		selectedMenuOption = OptionMenu.Start;
-		startLevel = new scenes.GameScene();
+		startLevel = new GameScene();
 		
 		keyboard.setBehavior(Keyboard.DOWN_KEY, Keyboard.DETECT_INITIAL_PRESS_ONLY);
 		keyboard.setBehavior(Keyboard.UP_KEY, Keyboard.DETECT_INITIAL_PRESS_ONLY);
